workspace/view/requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.14.0
mysqlclient==2.1.1
pillow==10.2.0
PyMySQL==1.1.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
